+++ src/proxy/proxy_main.py
@@ -14,7 +14,6 @@ from proxy.proxy_settings import settings
 from proxy.utils.printers import dbg, c, prettytable
 from proxy.utils.parsers import get_contact_info, get_mail_headers
 from proxy.utils.emails import dbgmail, sendmail, messageToSend, notifyHandshake
-from proxy.utils.cryptography import decryptusingsq
 from proxy.utils.hooks import cleanup
 
 
@@ -32,9 +31,7 @@ def init_lockfile():
     locktime = 0
     lockpid = None
 
-    while (
-        os.path.isfile(settings["lockfilepath"]) and locktime < settings["locktimeout"]
-    ):
+    while os.path.isfile(settings["lockfilepath"]) and locktime < settings["locktimeout"]:
         lock = open(settings["lockfilepath"], "r")
         lockpid = lock.read()
         lock.close()
@@ -59,20 +56,14 @@ def init_lockfile():
                 )
         else:
             dbg(
-                "Lockfile doesn't contain any numeric PID ["
-                + str(lockpid)
-                + "]. Removing file",
+                "Lockfile doesn't contain any numeric PID [" + str(lockpid) + "]. Removing file",
                 pub=False,
             )
             cleanup()
         locktime += 1
         sleep(1)
 
-    if (
-        os.path.isfile(settings["lockfilepath"])
-        and lockpid is not None
-        and lockpid.isdigit()
-    ):
+    if os.path.isfile(settings["lockfilepath"]) and lockpid is not None and lockpid.isdigit():
         lockpid = int(lockpid)
         if lockpid > 1:
             try:
@@ -180,9 +171,7 @@ def enable_dts(message):
 
     """
     global settings
-    dts = get_mail_headers(
-        message.inmail, "Disposition-Notification-To"
-    )  # Debug To Sender
+    dts = get_mail_headers(message.inmail, "Disposition-Notification-To")  # Debug To Sender
 
     if len(dts) > 0:
         addr = dts[0]
@@ -214,7 +203,7 @@ def init_workdir(message):
     if not os.path.exists(workdirpath):
         os.makedirs(workdirpath)
 
-    os.environ["HOME"] = workdirpath    
+    os.environ["HOME"] = workdirpath
     os.chdir(workdirpath)
 
     settings["work_dir"] = workdirpath
@@ -225,8 +214,8 @@ def init_workdir(message):
         # On windows, set the local app data folder to be the same as the workdir so the databases can be created correctly
         os.environ["LOCALAPPDATA"] = settings["work_dir"]
 
-    if not os.path.exists(settings['home']):
-        os.makedirs(settings['home'])
+    if not os.path.exists(settings["home"]):
+        os.makedirs(settings["home"])
 
 
 # ## Check if Sequoia-DB already exists, if not import keys later using planck ########################
@@ -237,9 +226,7 @@ def check_initial_import():
     Check if keys.db already exists, if not import keys later using planck
     """
 
-    keys_db_path = os.path.join(
-        os.environ["HOME"], settings["database_folder"], "keys.db"
-    )
+    keys_db_path = os.path.join(os.environ["HOME"], settings["database_folder"], "keys.db")
     return not os.path.exists(keys_db_path)
 
 
@@ -269,9 +256,7 @@ def load_planck():
 
     databases = ["management.db", "keys.db"]
     for database in databases:
-        db_path = os.path.join(
-            settings["work_dir"], settings["database_folder"], database
-        )
+        db_path = os.path.join(settings["work_dir"], settings["database_folder"], database)
         if not os.path.exists(db_path):
             blank_db_path = os.path.join(settings["project_root"], "data", database)
             shutil.copy(blank_db_path, db_path)
@@ -346,22 +331,16 @@ def create_planck_message(planck, message):
         from_username = inmail_parsed.from_.username
         from_address = inmail_parsed.from_.address
 
-        from_username_part = (
-            c(from_username, 2) if from_username and len(from_username) > 0 else ""
-        )
+        from_username_part = c(from_username, 2) if from_username and len(from_username) > 0 else ""
         from_address_part = c(" <" + from_address + ">", 3)
 
         to_username = inmail_parsed.to[0].username
         to_address = inmail_parsed.to[0].address
 
-        to_username_part = (
-            c(to_username, 2) if to_username and len(to_username) > 0 else ""
-        )
+        to_username_part = c(to_username, 2) if to_username and len(to_username) > 0 else ""
         to_address_part = c(" <" + to_address + ">", 3)
 
-        dbg(
-            f"Processing message from {from_username_part}{from_address_part} to {to_username_part}{to_address_part}"
-        )
+        dbg(f"Processing message from {from_username_part}{from_address_part} to {to_username_part}{to_address_part}")
 
     except Exception:
         e = sys.exc_info()
@@ -376,9 +355,7 @@ def create_planck_message(planck, message):
 
     # Log parsed message
 
-    logfilename = os.path.join(
-        settings["logpath"], "in." + settings["mode"] + ".parsed.eml"
-    )
+    logfilename = os.path.join(settings["logpath"], "in." + settings["mode"] + ".parsed.eml")
     dbg("planck-parsed message: " + c(logfilename, 6))
     logfile = codecs.open(logfilename, "w", "utf-8")
     logfile.write(str(inmail_parsed))
@@ -401,25 +378,9 @@ def process_message(planck, message):
     """
     try:
         if settings["mode"] == "decrypt":
-            # TODO: store some sort of failure-counter (per message ID?) to detect subsequent failures then
-            #  fallback to sq, then forward as-is
-            planckfails = False
-            if not planckfails:
-                dbg(c("Decrypting message via planck...", 2))
-                inmail_decrypted, keys, rating, flags = message.inmail_parsed.decrypt()
-                dbg(c("Decrypted in", 2), True)
-            else:
-                dbg(c("Decrypting message via Sequoia...", 2))
-                tmp = decryptusingsq(
-                    message.inmail,
-                    os.path.join(settings["keys_dir"], "sec.*.key"),
-                )
-                inmail_decrypted, keys, rating = (
-                    planck.Message(tmp[0]),
-                    tmp[1],
-                    None,
-                )
-                dbg(c("Decrypted in", 2), True)
+            dbg(c("Decrypting message via planck...", 2))
+            inmail_decrypted, keys, rating, flags = message.inmail_parsed.decrypt()
+            dbg(c("Decrypted in", 2), True)
 
             dbg(f"Message rating {rating}")
 
@@ -438,11 +399,7 @@ def process_message(planck, message):
                 if keys is None or len(keys) == 0:
                     dbg(c("Original message was NOT encrypted", 1))
                 else:
-                    dbg(
-                        c("Original message was encrypted to these keys", 2)
-                        + ":\n"
-                        + prettytable(list(set(keys)))
-                    )
+                    dbg(c("Original message was encrypted to these keys", 2) + ":\n" + prettytable(list(set(keys))))
 
         # Workaround for engine converting plaintext-only messages into a msg.txt inline-attachment
         # dst = str(dst).replace(' filename="msg.txt"', "")
@@ -463,15 +420,8 @@ def process_message(planck, message):
     message.inmail_decrypted = inmail_decrypted
 
     # Log processed message
-    logfilename = os.path.join(
-        settings["logpath"], "in." + settings["mode"] + ".processed.eml"
-    )
-    dbg(
-        "planck-processed message: "
-        + c(logfilename, 6)
-        + "\n"
-        + str(inmail_decrypted)[0:1337]
-    )
+    logfilename = os.path.join(settings["logpath"], "in." + settings["mode"] + ".processed.eml")
+    dbg("planck-processed message: " + c(logfilename, 6) + "\n" + str(inmail_decrypted)[0:1337])
     logfile = codecs.open(logfilename, "w", "utf-8")
     logfile.write(str(inmail_decrypted))
     logfile.close()
@@ -528,15 +478,9 @@ def filter_message(message):
 
         if settings["DEBUG"]:
             if stdout and len(stdout) > 0:
-                dbg(
-                    c("STDOUT:\n", 2)
-                    + prettytable(stdout.decode("utf8").strip().split("\n"))
-                )
+                dbg(c("STDOUT:\n", 2) + prettytable(stdout.decode("utf8").strip().split("\n")))
             if stderr and len(stderr) > 0:
-                dbg(
-                    c("STDERR:\n", 1)
-                    + prettytable(stderr.decode("utf8").strip().split("\n"))
-                )
+                dbg(c("STDERR:\n", 1) + prettytable(stderr.decode("utf8").strip().split("\n")))
             # dbg("Return code: " + c(str(rc), 3));
 
     dbg("Combined scan results:\n" + prettytable(scanresults))
@@ -544,11 +488,7 @@ def filter_message(message):
     if sum(scanresults.values()) == 0:
         dbg("All scans " + c("PASSED", 2) + ", relaying message", 2)
     else:
-        dbg(
-            "Some scans "
-            + c("FAILED", 1)
-            + ", not relaying message (keeping it in the Postfix queue for now)"
-        )
+        dbg("Some scans " + c("FAILED", 1) + ", not relaying message (keeping it in the Postfix queue for now)")
         admin_msg = f"A message from {message.msgfrom} and to {message.msgto} failed some of the scans."
         dbgmail(msg=admin_msg, subject="planck Proxy Scan failure")
         # sender_msg = f"Your message could not be delivered to {message.msg['msgto']}
@@ -571,20 +511,12 @@ def deliver_mail(message):
     dbg("Sending mail")
     dbg(
         "From: "
-        + (
-            (c(message.inmail_parsed.from_.username, 2))
-            if len(message.inmail_parsed.from_.username) > 0
-            else ""
-        )
+        + ((c(message.inmail_parsed.from_.username, 2)) if len(message.inmail_parsed.from_.username) > 0 else "")
         + c(" <" + message.inmail_parsed.from_.address + ">", 3)
     )
     dbg(
         "  To: "
-        + (
-            (c(message.inmail_parsed.to[0].username, 2))
-            if len(message.inmail_parsed.to[0].username) > 0
-            else ""
-        )
+        + ((c(message.inmail_parsed.to[0].username, 2)) if len(message.inmail_parsed.to[0].username) > 0 else "")
         + c(" <" + message.inmail_parsed.to[0].address + ">", 3)
     )